requirements.txt

streamlit>=1.36
streamlit-ace>=0.1.1
python-dotenv>=1.0.1
openai>=1.40.0
